% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/super.R
\name{multi_parent_env}
\alias{multi_parent_env}
\title{Hop over parent environments.}
\usage{
multi_parent_env(env, times)
}
\arguments{
\item{env}{environment}

\item{times}{integer}
}
\description{
Hop over parent environments.
}

