% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super-package.R
\docType{package}
\name{super-package}
\alias{super}
\alias{super-package}
\title{super: Interpreted String Literals}
\description{
An implementation of interpreted string literals. Based on the 'glue' package by Hester & Bryan (2024) \doi{10.32614/CRAN.package.glue} but with a focus on efficiency and simplicity at a cost of flexibility.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://timtaylor.codeberg.page/super/}
  \item Report bugs at \url{https://codeberg.org/TimTaylor/super/issues}
}

}
\author{
\strong{Maintainer}: Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID})

Authors:
\itemize{
  \item Jim Hester (\href{https://orcid.org/0000-0002-2739-7082}{ORCID})
  \item Jennifer Bryan (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
