% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/package.super.R
\docType{package}
\name{super}
\alias{super}
\alias{super-package}
\title{Giving R the ability to invoke "parent" methods
without the use of S3 generics.}
\description{
Giving R the ability to invoke "parent" methods
without the use of S3 generics.
}

