% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/package.super.R, R/super.R
\docType{package}
\name{super}
\alias{super}
\alias{super-package}
\title{Giving R the ability to invoke "parent" methods
without the use of S3 generics.}
\usage{
super(...)
}
\arguments{
\item{...}{The arguments passed to the parent function.}
}
\value{
the result of calling the next available function.
}
\description{
Giving R the ability to invoke "parent" methods
without the use of S3 generics.

Call the next available function by the same name.
}
\examples{
function1 <- function() {
  print("Top-level")
  invisible(NULL)
}

local({
  function2 <- function() {
    function1 <- function() {
      print("Mid-level")
      super::super()
    }

    function3 <- function() {
      function1 <- function() {
        print("Low-level")
        super::super()
      }
      function1()
    }

    function3()
  }

  function2()
})
# Will print
# [1] "Low-level"
# [1] "Mid-level"
# [1] "Top-level"
}

